+++ keycloak4s-admin-monix/src/main/scala/com/fullfacing/keycloak4s/admin/monix/client/KeycloakClient.scala
@@ -1,8 +1,6 @@
 package com.fullfacing.keycloak4s.admin.monix.client
 
-import cats.implicits._
 import com.fullfacing.keycloak4s.admin.client.{KeycloakClient => KeycloakClientA}
-import com.fullfacing.keycloak4s.admin.monix.utilities.ObservableExtensions.ObservableExtensions
 import com.fullfacing.keycloak4s.core.models._
 import monix.eval.Task
 import monix.reactive.Observable
@@ -23,29 +21,34 @@ class KeycloakClient[T](config: ConfigWithAuth)(implicit client: SttpBackend[Tas
    * @param batch  The amount of items each call should return.
    */
   def getList[A <: Any : Manifest](path: Seq[String], query: Seq[KeyValue] = Seq.empty[KeyValue], offset: Int = 0, limit: Int, batch: Int = 100): Observable[A] = {
-    val call = { i: Int =>
-      val max = if (i + batch >= limit) limit - i else batch
-      val q = (query :+ KeyValue("first", s"$i")) :+ KeyValue("max", s"$max")
-      get[Seq[A]](path, q).map(_.fold(throw _, res => res))
-    }
-
-    Observable.walk[Seq[A]](offset)(fetchResources(call, batch, limit)).flatMap(Observable.fromIterable)
+    Observable.unfoldEval(offset)(fetchResources(_, batch, limit, path, query)).flatMap(Observable.fromIterable)
   }
 
-  /** Created by https://github.com/Executioner1939
+  /**
+   * This function continually fetches the next set of resources until a result set less than the batch size is returned.
    *
-   * Generator functions for the `fromAsyncStateAction` on `Observable`. This function continually fetches
-   * the next set of resources until a result set less than the batch size is returned.
-   *
-   * @param source a function that fetches the next batch of resources.
    * @return the next state and element to be pushed downstream.
    */
-  def fetchResources[A](source: Int => Task[Seq[A]], batchSize: Int, limit: Int): Int => Task[Either[Seq[A], (Seq[A], Int)]] = { offset =>
-    source(offset).map { resources =>
-      if (resources.size >= batchSize && resources.size + offset < limit) {
-        (resources, offset + resources.size).asRight
-      } else {
-        resources.asLeft
+  def fetchResources[A <: Any : Manifest](offset: Int, batchSize: Int, limit: Int, path: Seq[String], query: Seq[KeyValue] = Seq.empty[KeyValue]): Task[Option[(Seq[A], Int)]] = {
+    if (offset == -1) { // Offset of -1 means we have nothing more to fetch
+      Task.now(None)
+
+    } else {
+      val q = (query :+ KeyValue("first", s"$offset")) :+ KeyValue("max", s"$batchSize")
+      get[Seq[A]](path, q).map(_.fold(throw _, res => res)).map { results =>
+
+        // Result set is less than batch size
+        if (results.size < batchSize) { // No more to fetch
+          Some((results, -1))
+
+        // Result set is equal to batch size
+        } else if (results.size == batchSize) { // More to Fetch
+          Some((results, offset + results.size))
+
+        // Result set is empty
+        } else {
+          None
+        }
       }
     }
   }