+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.1] - 2020-04-09
+### Changed
+- `Observable.walk` to use `Observable.unfoldEval`
+### Removed
+- `Observable.walk` extension function.
+- `AsyncState` used by `Observable.walk`
+
 ## [2.2.0] - 2020-04-08
 ### Added
 -  sbt release to handle automatic releases to Maven Central